[refactor]: Refactor water feature
